+++ Examples.BlazorServer/FileConverter.cs
@@ -1,0 +1,23 @@
+using Microsoft.AspNetCore.Components.Forms;
+using Microsoft.AspNetCore.Http;
+using System.IO;
+
+namespace Examples.BlazorServer;
+
+public static class WebFileConverter
+{
+    public static IFormFile ConvertToFormFile(this IBrowserFile browserFile)
+    {
+        ArgumentNullException.ThrowIfNull(browserFile);
+
+        var fileStream = new BrowserFileStream(browserFile);
+
+        var formFile = new FormFile(fileStream, 0, browserFile.Size, browserFile.Name, browserFile.Name)
+        {
+            Headers = new HeaderDictionary(),
+            ContentType = browserFile.ContentType
+        };
+
+        return formFile;
+    }
+}